# Conflicts:
#	hub/src/main/java/com/logistics/hub/application/dto/HubReadResDto.java
#	hub/src/main/java/com/logistics/hub/application/service/HubService.java
#	hub/src/main/java/com/logistics/hub/presentaion/controller/HubController.java
#	vendor/src/main/java/com/logistics/vendor/application/dto/VendorReadResDTO.java

+++ hub/src/main/java/com/logistics/hub/presentaion/controller/HubController.java
@@ -63,7 +63,6 @@ public class HubController {
 	@DeleteMapping("/hubs/{hubId}")
 	public ResponseEntity<String> deleteHub(
 		@AuthenticationPrincipal CustomPrincipal customPrincipal,
-		// @RequestHeader("X-User-id") String userId,
 		@PathVariable(name = "hubId") UUID hubId) {
 		hubService.deleteHub(hubId, customPrincipal.getUserId());
 		return ResponseEntity.ok("Hub successfully deleted");
@@ -91,10 +90,9 @@ public class HubController {
 	}
 
 	@GetMapping("/hubs/user/{userId}")
-	public UUID getUserHubId(@PathVariable("userId") int userId) {
+	public UUID getUserHubId(@PathVariable("userId") long userId) {
 		log.info("getUserHubId Controller");
 		return hubService.getUserHubId(userId);
 	}
-
 
 }
